database/requirements.txt
PyMySQL
python-dotenv
pandas